packages/manim-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
manim==0.18.0
python-dotenv==1.0.0
requests==2.31.0